+++ e3nn/nn/fc.py
@@ -1,12 +1,15 @@
 import torch
 
 from e3nn.math import normalize2mom
+from e3nn.util.jit import compile_mode
 
 
 def _identity(x):
     return x
 
 
+# This is a static network that can be traced
+@compile_mode('trace')
 class FullyConnectedNet(torch.nn.Module):
     r"""Fully-connected Neural Network
 
@@ -77,3 +80,10 @@ class FullyConnectedNet(torch.nn.Module):
                     x = x * self.variance_out**0.5
 
             return x
+
+    def _make_tracing_inputs(self, n: int = 1):
+        import random
+        return [
+            {'forward': (torch.randn(random.randint(1, 5), self.hs[0]),)}
+            for _ in range(n)
+        ]