+++ e3nn/util/test.py
@@ -5,6 +5,8 @@ import warnings
 import torch
 
 from e3nn import o3
+from e3nn.util.jit import compile, get_tracing_inputs, get_compile_mode
+from ._argtools import _get_args_in, _get_io_irreps, _transform
 
 
 FLOAT_TOLERANCE = {
@@ -90,7 +92,6 @@ def assert_equivariant(
     problems = {case: err for case, err in errors.items() if err > tolerance}
 
     if len(problems) != 0:
-        print(problems)
         errstr = (
             "Largest componentwise equivariance error was too large for: " + \
             '; '.join("(parity_k={:d}, did_translate={}) -> error={:.3e}".format(int(k[0]), bool(k[1]), float(v)) for k, v in problems.items())
@@ -195,36 +196,26 @@ def equivariance_error(
     return biggest_errs
 
 
-def assert_jit_trace(
+# TODO: this is only for things marked with @compile_mode.
+# Make something else for general script/tracability
+def assert_auto_jitable(
     func,
-    method_name=None,
-    args_in=None,
-    irreps_in=None,
-    irreps_out=None,
-    n_random_tests=2,
+    error_on_warnings=True,
+    n_trace_checks=2,
     strict_shapes=True,
-    **kwargs
 ):
-    r"""Assert that ``func`` can be traced to TorchScript.
+    r"""Assert that submodule ``func`` is automatically JITable.
 
     Parameters
     ----------
         func : Callable
             The function to trace.
-        method_name : str or None (default)
-            If ``func`` is a module, methods other than ``forward()`` can be traced by giving their name as a string. (This uses ``torch.jit.trace_module`` instead of ``torch.jit.trace``.)
-        args_in : list or None
-            the original input arguments for the function. If ``None`` and the function has ``irreps_in`` consisting only of ``o3.Irreps`` and ``'cartesian'``, random test inputs will be generated.
-        irreps_in : object
-            see ``equivariance_error``
-        irreps_out : object
-            see ``equivariance_error``
+        error_on_warnings : bool
+            If True (default), TracerWarnings emitted by ``torch.jit.trace`` will be treated as errors.
         n_random_tests : int
             If ``args_in`` is ``None`` and arguments are being automatically generated, this many random arguments will be generated as test inputs for ``torch.jit.trace``.
         strict_shapes : bool
             Test that the traced function errors on inputs with feature dimensions that don't match the input irreps.
-        **kwargs : kwargs
-            passed through to ``torch.jit.trace``.
     Returns
     -------
         The traced TorchScript function.
@@ -232,119 +223,40 @@ def assert_jit_trace(
     # Prevent pytest from showing this function in the traceback
     __tracebackhide__ = True
 
-    random_tests = args_in is None
-
-    args_in, irreps_in, irreps_out = _get_args_in(
-        func,
-        args_in=args_in,
-        irreps_in=irreps_in,
-        irreps_out=irreps_out
-    )
-
-    if random_tests:
-        test_inputs = [args_in] + [_rand_args(irreps_in) for _ in range(n_random_tests)]
-    else:
-        test_inputs = [args_in]
+    if get_compile_mode(func) is None:
+        raise ValueError("assert_auto_jitable is only for modules marked with @compile_mode")
 
     # Test tracing
     with warnings.catch_warnings():
-        warnings.filterwarnings('error', category=torch.jit.TracerWarning)
-        if method_name is not None:
-            func_trace = torch.jit.trace_module(
-                func,
-                inputs={method_name: tuple(args_in)},
-                check_inputs=[{method_name: t} for t in test_inputs]
-            )
-            func_trace = getattr(func_trace, method_name)
-        else:
-            func_trace = torch.jit.trace(
-                func,
-                example_inputs=tuple(args_in),
-                check_inputs=test_inputs
-            )
+        if error_on_warnings:
+            warnings.filterwarnings('error', category=torch.jit.TracerWarning)
+        func_jit = compile(
+            func,
+            n_trace_checks=n_trace_checks
+        )
 
     # Confirm that it rejects incorrect shapes
-    if random_tests and strict_shapes:
-        bad_args = _rand_args(irreps_in)
-        # Since _rand_args is OK, they're all Irreps style args where changing the feature dimension is wrong
-        bad_which = random.randint(0, len(bad_args)-1)
-        bad_args[bad_which] = bad_args[bad_which][..., :-random.randint(1, 3)]  # make bad shape
+    if strict_shapes:
         try:
-            func_trace(*bad_args)
-        except torch.jit.Error as e:
-            # As far as I can tell, there's no good way to introspect TorchScript exceptions. Checking for RuntimeError at least eliminates particlar possibilities
-            assert "RuntimeError" in str(e), "TorchScript through an unexpectedly strange error"
+            all_bad_args = get_tracing_inputs(func, n=1)[0]
+        except ValueError:
+            # couldn't infer, don't check
+            pass
         else:
-            raise AssertionError("Traced function didn't error on bad input shape")
+            for method, bad_args in all_bad_args.items():
+                # Since _rand_args is OK, they're all Irreps style args where changing the feature dimension is wrong
+                bad_which = random.randint(0, len(bad_args)-1)
+                bad_args = list(bad_args)
+                bad_args[bad_which] = bad_args[bad_which][..., :-random.randint(1, 3)]  # make bad shape
+                try:
+                    if method == 'forward':
+                        func_jit(*bad_args)
+                    else:
+                        getattr(func_jit, method)(*bad_args)
+                except (torch.jit.Error, RuntimeError):
+                    # As far as I can tell, there's no good way to introspect TorchScript exceptions.
+                    pass
+                else:
+                    raise AssertionError("Traced function didn't error on bad input shape")
 
-    return func_trace
-
-
-def _transform(dat, irreps_dat, rot_mat, translation=0.):
-    """Transform ``dat`` by ``rot_mat`` and ``translation`` according to ``irreps_dat``."""
-    out = []
-    for irreps, a in zip(irreps_dat, dat):
-        if irreps is None:
-            out.append(a)
-        elif irreps == 'cartesian_points':
-            out.append((a @ rot_mat.T) + translation)
-        elif irreps == 'cartesian_vectors':
-            out.append((a @ rot_mat.T))
-        else:
-            # For o3.Irreps
-            out.append(a @ irreps.D_from_matrix(rot_mat).T)
-    return out
-
-
-def _get_io_irreps(func, irreps_in=None, irreps_out=None):
-    """Preprocess or, if not given, try to infer the I/O irreps for ``func``."""
-    SPECIAL_VALS = ['cartesian_points', 'cartesian_vectors', None]
-
-    if irreps_in is None:
-        if hasattr(func, 'irreps_in'):
-            irreps_in = [func.irreps_in]
-        elif hasattr(func, 'irreps_in1'):
-            irreps_in = [func.irreps_in1, func.irreps_in2]
-        else:
-            raise ValueError("Cannot infer irreps_in for %r; provide them explicitly" % func)
-    if irreps_out is None:
-        if hasattr(func, 'irreps_out'):
-            irreps_out = [func.irreps_out]
-        else:
-            raise ValueError("Cannot infer irreps_out for %r; provide them explicitly" % func)
-
-    if isinstance(irreps_in, o3.Irreps) or irreps_in in SPECIAL_VALS:
-        irreps_in = [irreps_in]
-    elif isinstance(irreps_in, list) or isinstance(irreps_in, tuple):
-        irreps_in = [i if i in SPECIAL_VALS else o3.Irreps(i) for i in irreps_in]
-    else:
-        irreps_in = [o3.Irreps(irreps_in)]
-
-    if isinstance(irreps_out, o3.Irreps) or irreps_out in SPECIAL_VALS:
-        irreps_out = [irreps_out]
-    elif isinstance(irreps_out, list) or isinstance(irreps_out, tuple):
-        irreps_out = [i if i in SPECIAL_VALS else o3.Irreps(i) for i in irreps_out]
-    else:
-        irreps_out = [o3.Irreps(irreps_out)]
-
-    return irreps_in, irreps_out
-
-
-def _get_args_in(func, args_in=None, irreps_in=None, irreps_out=None):
-    irreps_in, irreps_out = _get_io_irreps(func, irreps_in=irreps_in, irreps_out=irreps_out)
-    if args_in is None:
-        args_in = _rand_args(irreps_in)
-    assert len(args_in) == len(irreps_in), "irreps_in and args_in don't match in length"
-    return args_in, irreps_in, irreps_out
-
-
-def _rand_args(irreps_in):
-    if not all((isinstance(i, o3.Irreps) or i == 'cartesian_points') for i in irreps_in):
-        raise ValueError("Random arguments cannot be generated when argument types besides Irreps and `'cartesian_points'` are specified; provide explicit ``args_in``")
-    # Generate random args with random size batch dim between 1 and 4:
-    batch_size = random.randint(1, 4)
-    args_in = [
-        torch.randn(batch_size, 3) if (irreps == 'cartesian_points') else irreps.randn(batch_size, -1)
-        for irreps in irreps_in
-    ]
-    return args_in
+    return func_jit